pin view refactor
